# model/swagger.yml
openapi: 3.0.0
info:
  description: |
    Registry API enabling advanced search on PDS data and metadata. The API provides end-points to search for bundles, collections and any PDS products with advanced search queries. It also enables to browse the archive hierarchically downward (e.g. collection/s products) or upward (e.g. bundles containing a product).
    The detailed syntax for querying the end-point is given in the reference documentation.
  version: 1.0.0-SNAPSHOT
  title: Registry API
  termsOfService: 'http://pds.nasa.gov'
  contact:
    email: pds-operator@jpl.nasa.gov
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: https://pds.nasa.gov/api/registry/1.0
    description: production server
...
paths:
  /products:
    get:
      tags:
        - products
      summary: |
        returns all PDS data products, including bundles, collections, documentation, context and observational products.
      operationId: products
      responses:
        '200':
          $ref: "#/components/responses/Plural"
        '400':
          $ref: "#/components/responses/Error"
        '404':
          $ref: "#/components/responses/Error"
        '500':
          $ref: "#/components/responses/Error"
        '501':
          $ref: "#/components/responses/Error"
      parameters:
        - $ref: "#/components/parameters/Fields"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/Sort"
        - $ref: "#/components/parameters/Start"
        - $ref: "#/components/parameters/SummaryOnly"
...
components:
  parameters:
    Fields:
      name: fields
      in: query
      description: |
        syntax: fields=field1,field2,...

        behavior: this parameter and the headder Accept: type determine what content is packaged for the result. While the types application/csv, application/kvp+json, and text/csv return only the fields requesteted, all of the other types have a minimal set of fields that must be returned. Duplicating a minimally required field in this parameter has not effect. The types vnd.nasa.pds.pds4+json and vnd.nasa.pds.pds4+xml have a complete set of fields that must be returned; meaning this parameter does not impact their content. When fields is not used, then the minimal set of fields, or all when minimal is an empty set, is returned.

        notes: the blob fields are blocked unless specifically requrested and only for the */csv and application/kvp+csv types.
      required: false
      schema:
        type: array
        items:
          type: string
...
    Keyword:
      name: keywords
      in: query
      description: |
        syntax: keyword=keyword1,keyword2,...

        behavior: free text search on title and description (if set q is ignored

        notes: is this implemented?
      required: false
      schema:
        type: array
        items:
          type: string
    Limit:
      name: limit
      in: query
      description: |
        syntax: limit=10

        behavior: maximum number of matching results returned, for pagination
      required: false
      schema:
        type: integer
        minimum: 1
        default: 100
    Query:
      name: q
      in: query
      description: |
        syntax: q="vid eq 13.0"

        behavior: query uses eq,ne,gt,ge,lt,le,(,),not,and,or operators. Properties are named as in 'properties' attributes, literals are strings between quotes, like "animal", or numbers. Detailed query specification is available at https://bit.ly/3h3D54T

        note: ignored when keyword is present
      required: false
      schema:
        type: string
    Sort:
      name: sort
      in: query
      description: |
        syntax: sort=asc(field0),desc(field1),...

        behavior: is this implemented?
      required: false
      schema:
        type: array
        items:
          type: string
    Start:
      name: start
      in: query
      description: |
        syntax: start=12

        behavior: offset in matching result list, for pagination
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    SummaryOnly:
      name: summary-only
      in: query
      description: |
        syntax: summary-only={true,false}

        behavior: only return the summary when a list is returned. Useful to get the list of available properties.
      schema:
        type: boolean
        default: false
  responses:
    Error:
      description: Unsuccessful request
      content:
        "*":
          schema:
            type: object
        "*/*":
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/csv:
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/json:
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/kvp+json:
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/vnd.nasa.pds.pds4+json:
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/vnd.nasa.pds.pds4+xml:
          schema:
            $ref: '#/components/schemas/errorMessage'
        application/xml:
          schema:
            $ref: '#/components/schemas/errorMessage'
        text/csv:
          schema:
            $ref: '#/components/schemas/errorMessage'
        text/html:
          schema:
            $ref: '#/components/schemas/errorMessage'
        text/xml:
          schema:
            $ref: '#/components/schemas/errorMessage'
    Plural:
      description: Successful request
      content:
        "*":
          schema:
            type: object
        "*/*":
          schema:
            $ref: '#/components/schemas/pdsProducts'
        application/csv:
          schema:
            $ref: '#/components/schemas/wyriwygProducts'
        application/json:
          schema:
            $ref: '#/components/schemas/pdsProducts'
        application/kvp+json:
          schema:
            $ref: '#/components/schemas/wyriwygProducts'
        application/vnd.nasa.pds.pds4+json:
          schema:
            $ref: '#/components/schemas/pds4Products'
        application/vnd.nasa.pds.pds4+xml:
          schema:
            $ref: '#/components/schemas/pds4Products'
        application/xml:
          schema:
            $ref: '#/components/schemas/pdsProducts'
        text/csv:
          schema:
            $ref: '#/components/schemas/wyriwygProducts'
        text/html:
          schema:
            $ref: '#/components/schemas/pdsProducts'
        text/xml:
          schema:
            $ref: '#/components/schemas/pdsProducts'
...
  schemas:
    summary:
      type: object
      properties:
        q:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        hits:
          description: total number of matching results. -1 indicates an excessive amount of resource required to compute it
          type: integer
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        took:
          description: Number of milliseconds it "took" to do request
          type: integer
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        start:
          type: integer
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        limit:
          type: integer
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        sort:
          type: array
          items:
            type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        properties:
          type: array
          items:
            type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
      xml:
        prefix: 'pds_api'
        namespace: 'http://pds.nasa.gov/api'
    pdsProduct:
      type: object
      required:
       - id
       - metadata
      properties:
        id:
          type: string
          description: identifier lidvid of the collection
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        type:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        title:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        description:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        start_date_time:
          type: string
          description: start date time of the observations in ISO8601
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        stop_date_time:
          type: string
          description: stop date time of the observations in ISO8601
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        investigations:
          type: array
          description: list of missions or observing campaigns which produced the data
          items:
            $ref: '#/components/schemas/reference'
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        observing_system_components: 
          type: array
          description: list of instruments or platforms generating the data
          items:
            $ref: '#/components/schemas/reference'
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        targets: 
          type: array
          description: list of targets or feature of interest the observation.
          items:
            $ref: '#/components/schemas/reference'
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        metadata:
          $ref: '#/components/schemas/metadata'
        properties:
          type: object
          description: properties extracted from pds4 model, syntax for property name is <namespace>:<property>[/<namespace>:<property>]*
          additionalProperties:
            $ref: '#/components/schemas/propertyArrayValues'
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
      xml:
        prefix: 'pds_api'
        namespace: 'http://pds.nasa.gov/api'
    pdsProducts:
      type: object
      required:
        - summary
        - data
      properties:
        summary:
          $ref: '#/components/schemas/summary'
        data:
          items:
            $ref: '#/components/schemas/pdsProduct'
    wyriwygProduct:
      type: object
      required:
        - key_value_pairs
      properties:
        key_value_pairs:
          type: array
          items:
            type: object
            properties:
              key:
                type: string
              value:
                type: string
    wyriwygProducts:
      type: object
      required:
        - summary
        - data
      properties:
        summary:
          $ref: '#/components/schemas/summary'
        data:
          items:
            $ref: '#/components/schemas/wyriwygProduct'
    propertyArrayValues:
      type: array 
      items:
        type: string
      xml:
        prefix: 'pds_api'
        namespace: 'http://pds.nasa.gov/api'
...
    errorMessage:
      type: object
      required:
        - request
        - message
      properties:
        request:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
        message:
          type: string
          xml:
            prefix: 'pds_api'
            namespace: 'http://pds.nasa.gov/api'
      xml:
        prefix: 'pds_api'
        namespace: 'http://pds.nasa.gov/api'
      example:
        {
          request: 'https://pds.nasa.gov/api/1.0/collections?q=mission qt 12',
          message: 'query operator gt not supported on field mission'
        }
...
    pds4Products:
      type: object
      required:
        - summary
        - data
      properties:
        summary:
          $ref: '#/components/schemas/summary'
        data:
          items:
            $ref: '#/components/schemas/pds4Product'
...